fastapi/requirements.txt
fastapi
gunicorn
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary